requirements.txt
Faker==0.8.8
pytest==3.3.1
pytz==2017.3
selenium==3.8.1
Appium_Python_Client==0.25
